+++ flare_dart/lib/actor_color.dart
@@ -1,15 +1,14 @@
+import "dart:collection";
 import "dart:typed_data";
-import 'math/mat2d.dart';
 
+import "actor_artboard.dart";
+import "actor_component.dart";
+import "actor_flags.dart";
 import "actor_node.dart";
 import "actor_shape.dart";
-import "actor_artboard.dart";
-
-import "actor_component.dart";
-import "dart:collection";
-import "stream_reader.dart";
+import 'math/mat2d.dart';
 import "math/vec2d.dart";
-import "actor_flags.dart";
+import "stream_reader.dart";
 
 enum FillRule { EvenOdd, NonZero }
 enum StrokeCap { Butt, Round, Square }
@@ -53,6 +52,7 @@ abstract class ActorPaint extends ActorComponent {
     return component;
   }
 
+  @override
   void completeResolve() {
     artboard.addDependency(this, parent);
   }
@@ -221,6 +221,7 @@ abstract class ColorFill extends ActorColor with ActorFill {
     return component;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();
 
@@ -244,6 +245,7 @@ abstract class ColorStroke extends ActorColor with ActorStroke {
     return component;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();
 
@@ -256,11 +258,12 @@ abstract class ColorStroke extends ActorColor with ActorStroke {
 
 abstract class GradientColor extends ActorPaint {
   Float32List _colorStops = Float32List(10);
-  Vec2D _start = Vec2D();
-  Vec2D _end = Vec2D();
-  Vec2D _renderStart = Vec2D();
-  Vec2D _renderEnd = Vec2D();
-  double opacity = 1.0;
+  final Vec2D _start = Vec2D();
+  final Vec2D _end = Vec2D();
+  final Vec2D _renderStart = Vec2D();
+  final Vec2D _renderEnd = Vec2D();
+  @override
+  final double opacity = 1.0;
 
   Vec2D get start => _start;
   Vec2D get end => _end;
@@ -315,6 +318,7 @@ abstract class GradientFill extends GradientColor with ActorFill {
     return component;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();
 
@@ -338,6 +342,7 @@ abstract class GradientStroke extends GradientColor with ActorStroke {
     return component;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();
 
@@ -381,6 +386,7 @@ abstract class RadialGradientFill extends RadialGradientColor with ActorFill {
     return component;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();
 
@@ -406,6 +412,7 @@ abstract class RadialGradientStroke extends RadialGradientColor
     return component;
   }
 
+  @override
   void completeResolve() {
     super.completeResolve();