+++ flare_flutter/lib/flare.dart
@@ -290,11 +290,11 @@ class FlutterColorFill extends ColorFill with FlutterFill {
 
   Color get uiColor {
     Float32List c = displayColor;
-    return Color.fromRGBO(
-        (c[0] * 255.0).round(),
-        (c[1] * 255.0).round(),
-        (c[2] * 255.0).round(),
-        c[3] * artboard.modulateOpacity * opacity * shape.renderOpacity);
+    double o = (artboard.modulateOpacity * opacity * shape.renderOpacity)
+        .clamp(0.0, 1.0)
+        .toDouble();
+    return Color.fromRGBO((c[0] * 255.0).round(), (c[1] * 255.0).round(),
+        (c[2] * 255.0).round(), c[3] * o);
   }
 
   set uiColor(Color c) {
@@ -322,11 +322,11 @@ class FlutterColorStroke extends ColorStroke with FlutterStroke {
 
   Color get uiColor {
     Float32List c = displayColor;
-    return Color.fromRGBO(
-        (c[0] * 255.0).round(),
-        (c[1] * 255.0).round(),
-        (c[2] * 255.0).round(),
-        c[3] * artboard.modulateOpacity * opacity * shape.renderOpacity);
+    double o = (artboard.modulateOpacity * opacity * shape.renderOpacity)
+        .clamp(0.0, 1.0)
+        .toDouble();
+    return Color.fromRGBO((c[0] * 255.0).round(), (c[1] * 255.0).round(),
+        (c[2] * 255.0).round(), c[3] * o);
   }
 
   set uiColor(Color c) {
@@ -355,11 +355,12 @@ class FlutterGradientFill extends GradientFill with FlutterFill {
 
     int idx = 0;
     for (int i = 0; i < numStops; i++) {
+      double o = colorStops[idx + 3].clamp(0.0, 1.0).toDouble();
       ui.Color color = ui.Color.fromRGBO(
           (colorStops[idx] * 255.0).round(),
           (colorStops[idx + 1] * 255.0).round(),
           (colorStops[idx + 2] * 255.0).round(),
-          colorStops[idx + 3]);
+          o);
       colors.add(color);
       stops.add(colorStops[idx + 4]);
       idx += 5;
@@ -373,14 +374,17 @@ class FlutterGradientFill extends GradientFill with FlutterFill {
               .toDouble());
     } else {
       Float32List overrideColor = artboard.overrideColor;
+      double o = (overrideColor[3] *
+              artboard.modulateOpacity *
+              opacity *
+              shape.renderOpacity)
+          .clamp(0.0, 1.0)
+          .toDouble();
       paintColor = ui.Color.fromRGBO(
           (overrideColor[0] * 255.0).round(),
           (overrideColor[1] * 255.0).round(),
           (overrideColor[2] * 255.0).round(),
-          overrideColor[3] *
-              artboard.modulateOpacity *
-              opacity *
-              shape.renderOpacity);
+          o);
     }
     _paint
       ..color = paintColor
@@ -408,11 +412,12 @@ class FlutterGradientStroke extends GradientStroke with FlutterStroke {
 
     int idx = 0;
     for (int i = 0; i < numStops; i++) {
+      double o = colorStops[idx + 3].clamp(0.0, 1.0).toDouble();
       ui.Color color = ui.Color.fromRGBO(
           (colorStops[idx] * 255.0).round(),
           (colorStops[idx + 1] * 255.0).round(),
           (colorStops[idx + 2] * 255.0).round(),
-          colorStops[idx + 3]);
+          o);
       colors.add(color);
       stops.add(colorStops[idx + 4]);
       idx += 5;
@@ -426,14 +431,17 @@ class FlutterGradientStroke extends GradientStroke with FlutterStroke {
               .toDouble());
     } else {
       Float32List overrideColor = artboard.overrideColor;
+      double o = (overrideColor[3] *
+              artboard.modulateOpacity *
+              opacity *
+              shape.renderOpacity)
+          .clamp(0.0, 1.0)
+          .toDouble();
       paintColor = ui.Color.fromRGBO(
           (overrideColor[0] * 255.0).round(),
           (overrideColor[1] * 255.0).round(),
           (overrideColor[2] * 255.0).round(),
-          overrideColor[3] *
-              artboard.modulateOpacity *
-              opacity *
-              shape.renderOpacity);
+          o);
     }
     _paint
       ..color = paintColor
@@ -463,11 +471,12 @@ class FlutterRadialFill extends RadialGradientFill with FlutterFill {
 
     int idx = 0;
     for (int i = 0; i < numStops; i++) {
+      double o = colorStops[idx + 3].clamp(0.0, 1.0).toDouble();
       ui.Color color = ui.Color.fromRGBO(
           (colorStops[idx] * 255.0).round(),
           (colorStops[idx + 1] * 255.0).round(),
           (colorStops[idx + 2] * 255.0).round(),
-          colorStops[idx + 3]);
+          o);
       colors.add(color);
       stops.add(colorStops[idx + 4]);
       idx += 5;
@@ -487,14 +496,17 @@ class FlutterRadialFill extends RadialGradientFill with FlutterFill {
               .toDouble());
     } else {
       Float32List overrideColor = artboard.overrideColor;
+      double o = (overrideColor[3] *
+              artboard.modulateOpacity *
+              opacity *
+              shape.renderOpacity)
+          .clamp(0.0, 1.0)
+          .toDouble();
       paintColor = ui.Color.fromRGBO(
           (overrideColor[0] * 255.0).round(),
           (overrideColor[1] * 255.0).round(),
           (overrideColor[2] * 255.0).round(),
-          overrideColor[3] *
-              artboard.modulateOpacity *
-              opacity *
-              shape.renderOpacity);
+          o);
     }
 
     _paint
@@ -523,11 +535,12 @@ class FlutterRadialStroke extends RadialGradientStroke with FlutterStroke {
 
     int idx = 0;
     for (int i = 0; i < numStops; i++) {
+      double o = colorStops[idx + 3].clamp(0.0, 1.0).toDouble();
       ui.Color color = ui.Color.fromRGBO(
           (colorStops[idx] * 255.0).round(),
           (colorStops[idx + 1] * 255.0).round(),
           (colorStops[idx + 2] * 255.0).round(),
-          colorStops[idx + 3]);
+          o);
       colors.add(color);
       stops.add(colorStops[idx + 4]);
       idx += 5;
@@ -541,14 +554,17 @@ class FlutterRadialStroke extends RadialGradientStroke with FlutterStroke {
               .toDouble());
     } else {
       Float32List overrideColor = artboard.overrideColor;
+      double o = (overrideColor[3] *
+              artboard.modulateOpacity *
+              opacity *
+              shape.renderOpacity)
+          .clamp(0.0, 1.0)
+          .toDouble();
       paintColor = ui.Color.fromRGBO(
           (overrideColor[0] * 255.0).round(),
           (overrideColor[1] * 255.0).round(),
           (overrideColor[2] * 255.0).round(),
-          overrideColor[3] *
-              artboard.modulateOpacity *
-              opacity *
-              shape.renderOpacity);
+          o);
     }
 
     _paint
@@ -706,6 +722,11 @@ class FlutterActorArtboard extends ActorArtboard {
       : super(actor, isInstance);
 
   void draw(ui.Canvas canvas) {
+    if(clipContents) {
+      canvas.save();
+      AABB aabb = artboardAABB();
+      canvas.clipRect(Rect.fromLTRB(aabb[0], aabb[1], aabb[2], aabb[3]));
+    }
     for (final ActorDrawable drawable in drawableNodes) {
       if (drawable is FlutterActorDrawable) {
         (drawable as FlutterActorDrawable).draw(canvas);
@@ -728,6 +749,9 @@ class FlutterActorArtboard extends ActorArtboard {
             Offset(start[0], start[1]), Offset(end[0], end[1]), debugPaint);
       }
     }*/
+    if(clipContents) {
+      canvas.restore();
+    }
   }
 
   void dispose() {}