+++ flare_flutter/CHANGELOG.md
@@ -1,3 +1,17 @@
+## [1.5.7] - 2019-08-26 10:39:49
+
+- Clip the artboard based on settings from Flare.
+
+## [1.5.6] - 2019-08-23 09:52:13
+
+- Clamp opacity values into 0..1 range before creating color for paint.
+
+## [1.5.5] - 2019-07-24 11:44:36
+
+ - Adding artboard option to FlareActor. Use this to change which artboard gets displayed by the FlareActor widget.
+ - Fixed incorrect signature of load method. If you were deriving FlareRenderBox, you'll need to update it to match. It's a minor change from void to Future<void>.
+ - Added some documentation to the FlareActor parameters.
+
 ## [1.5.4] - 2019-07-08 21:10:50
 
  - Using Uint16List for vertex indices now that Flutter Stable has been updated.